Q4 parte2

+++ questao4.c
@@ -2,7 +2,7 @@
 #include <limits.h>
 
 void main (){
-    int minimo = INT_MIN;
-    int maximo = INT_MAX;
-    printf("menor: %d\nmaior: %d", minimo, maximo);
+    int minimo = SHRT_MIN;
+    long long int maximo = LLONG_MAX;
+    printf("menor: %d\nmaior: %lld", minimo, maximo);
 }
